merge module_create and module_load
the original intention of the separation was to allow using
statically allocated struct modules. however, it doesn't make much
sense as we have unexported module_unload since then.
while it might or might not make sense to restore module_unload,
unify module_create and module_load for now so that it's at least
consistent.